remove the worker process.
Things seem to be going fast enough in a desktop to skip the need for using a Worker Proces for now.
Tweaked the event listeners to reuse a little more code, and clear the faded cards a little more often because of some problems i found with the way I am marking the cards up to now.